Moved: Color565 de/recorrelate tests to proper module
